Ali-Alfaqih/requirements.txt
torch>=2.6.0
torchvision>=0.17.0
numpy>=1.24.3
matplotlib>=3.7.1
scikit-learn>=1.2.2
flask>=2.3.2
pillow>=9.5.0
tqdm>=4.65.0
ipython>=8.13.2
jupyter>=1.0.0